% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directeffect-package.R
\docType{package}
\name{directeffect-package}
\alias{directeffect}
\alias{directeffect-package}
\title{directeffect: Estimate Latent Direct Drug Effects from Networks of Comparative Estimates}
\description{
Estimates latent direct drug effects from a network of comparative-effect estimates. Comparative estimates identify only differences between drugs (the effect surface); absolute placebo anchors position that surface against the placebo-defined sea level. Provides frequentist and Bayesian estimation engines behind a single interface, together with diagnostics for network coherence and identifiability.
}
\author{
\strong{Maintainer}: Adam Black \email{ablack3@gmail.com}

}
\keyword{internal}
